+++ doc/changelog.md
@@ -10,7 +10,9 @@
 ### Configs
 * made Roadroller paving/destruction configurable via server config
 * C-Moon's Ult range is now configurable via its moveset
+* C-Moon's Ult can exclude entities via "gravitychanger:forbidden_entities" and "gravitychanger:forbidden_entity_rendering"; this is not new, but now item frames and paintings are included by default
 * D4C's Dimensional Hop duration is now configurable via its moveset
+* D4C's Dimensional Hop can now have blacklisted blocks via the block tag "jcraft:au_replaced_with_air"
 * Mandom, BTD Ult and GER RTZ max reach is configurable via their movesets
 ### Commands
 * …